// src/components/TableComponents/StandardTableTypes.d.ts
export type ConfigItemType<DataItem, StyleGeneric> = {
  /** Unique name for the column. Should be associated with the name that you are sorting on for the column. */
  name: SortColumnProps['propName']
  /** Unique label for the column headers. */
  label: SortColumnProps['label']
  /** Optionally add custom JSX. This will be rendered to the right of the column header. */
  columnHeaderSubContent?: React.ReactNode
  /** Used if you have multiple options to sort on within a column. */
  options?: SortColumnProps['options']
  /** The JSX that will be rendered in a cell. */
  cell: {
    className?: string | ((data: DataItem) => string)
    children: (data: DataItem, index?: number) => React.ReactNode
    /** Optional style overrides for a cell. */
    style?: StyleGeneric
  }
  /** Determines the static first column in mobile and also the column that will not be able to be removed in column customizations. */
  mainColumn?: boolean
  /** Used if the API requires a lowercase parameter to be passed in. */
  lowerCaseParam?: SortColumnProps['lowerCaseParam']
  /** Define the tooltip for a column. */
  tooltip?: SortColumnProps['tooltip']
  /** Class name for a column. */
  className?: string
  /** Prevent sorting for a column. */
  noSort?: boolean
  /** Column that has a single button. Used for display purposes in mobile. */
  isButton?: boolean
  /** Filter for a column. */
  filter?: SortColumnProps['filter']
  /** Style overrides for a column. */
  style?: StyleGeneric
}